Контроллеры не являются внешними вызовами. Внешний вызов = вызов из приложения вовне.
