--- spring-18/library/src/main/resources/templates/waitPage.html
@@ -1,15 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" xmlns:th="http://www.thymeleaf.org">
-<head>
-    <title>Wait Page</title>
-</head>
-
-<body>
-<h1>Sorry for your wait !</h1>
-<button>
-    <thead>
-    <a th:href="@{/book}" href="booklist.html">Back</a>
-    </thead>
-</button>
-</body>
-</html>
